+++ cms/src/app/App.tsx
@@ -3,19 +3,23 @@ import { RouterProvider } from "@tanstack/react-router";
 import AuthProvider, { useAuth } from "../hooks/useAuth";
 import UserProvider from "../hooks/useUser";
 import router from "../config/router.config";
-import React, { Children, useCallback, useState } from "react";
-import useAxiosPrivateConfig from "../hooks/useAxiosPrivate";
+import React, { useCallback, useState } from "react";
+import AxiosPrivateProvider, {
+  useAxiosPrivate,
+  useAxiosPrivateConfig,
+} from "../hooks/useAxiosPrivate";
 
 const InnerApp = () => {
   console.group("InnerApp running...");
   const auth = useAuth();
+  const axiosPrivate = useAxiosPrivate();
   console.log("auth:", auth);
   console.groupEnd();
   if (auth.isLoading) {
     return <p>Loading...</p>;
   }
 
-  return <RouterProvider router={router} context={{ auth }} />;
+  return <RouterProvider router={router} context={{ auth, axiosPrivate }} />;
 };
 
 const AppProviders: React.FC<{ children: React.ReactNode }> = ({
@@ -23,15 +27,13 @@ const AppProviders: React.FC<{ children: React.ReactNode }> = ({
 }) => {
   console.log("[AppProviders] rendering...");
   const [accessToken, setAccessToken] = useState<string | null>(null);
-  const getAccessToken = useCallback(() => accessToken, [accessToken]);
+  // const getAccessToken = useCallback(() => accessToken, [accessToken]);
   const updateAccessToken = useCallback((newAccessToken: string | null) => {
     setAccessToken(newAccessToken);
   }, []);
 
   const axiosPrivate = useAxiosPrivateConfig({
     accessToken,
-    setAccessToken,
-    getAccessToken,
     updateAccessToken,
   });
 
@@ -42,7 +44,9 @@ const AppProviders: React.FC<{ children: React.ReactNode }> = ({
         updateAccessToken={updateAccessToken}
         axiosPrivate={axiosPrivate}
       >
-        <UserProvider>{children}</UserProvider>
+        <AxiosPrivateProvider axiosPrivate={axiosPrivate}>
+          <UserProvider>{children}</UserProvider>
+        </AxiosPrivateProvider>
       </AuthProvider>
     </>
   );
@@ -51,21 +55,11 @@ const AppProviders: React.FC<{ children: React.ReactNode }> = ({
 const App = () => {
   return (
     <div id="app">
-      <AuthProvider>
-        <UserProvider>
-          <InnerApp />
-        </UserProvider>
-      </AuthProvider>
+      <AppProviders>
+        <InnerApp />
+      </AppProviders>
     </div>
   );
-
-  // return (
-  //   <div id="app">
-  //     <AppProviders>
-  //       <InnerApp />
-  //     </AppProviders>
-  //   </div>
-  // );
 };
 
 export default App;