+++ cms/src/routes/_protected/_dashboard/_user.tsx
@@ -1,9 +1,10 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_protected/_dashboard/_user")({
-  loader: async ({ params }) => {
+  loader: async ({ params, context }) => {
     console.log("[/_user] loader running...");
     console.log("params:", params);
+    console.log("context:", context);
   },
   component: RouteComponent,
 });