+++ cms/src/routes/_protected/_dashboard/_user/$username/posts.tsx
@@ -3,6 +3,14 @@ import { createFileRoute, useRouteContext } from "@tanstack/react-router";
 export const Route = createFileRoute(
   "/_protected/_dashboard/_user/$username/posts"
 )({
+  loader: async ({ context, params }) => {
+    console.group("[/posts] loader running...");
+    console.log("context:", context);
+    console.log("params:", params);
+    console.groupEnd();
+    const response = await context.axiosPrivate.get(`/users`);
+    console.log("response:", response);
+  },
   component: RouteComponent,
 });